moved some files from components to pages, added new user card component and finished basic styling for feed page
