+++ client/src/App.js
@@ -1,16 +1,17 @@
 import React, { Component } from "react";
 // import { BrowserRouter as Router, Route } from "react-router-dom";
 import { Switch, Route, Link, BrowserRouter as Router  } from "react-router-dom";
+
 // import Container from "../src/components/Container";
 // import Brew from "../src/pages/Brew";
 // import Profile from "../src/pages/Profile";
-// import Feed from "../src/pages/Feed";
 // import Footer from "../src/components/Footer";
 // import logo from './logo.svg';
 import './App.css';
 
 import AuthService from "./services/auth.service";
 
+import Feed from "../src/pages/Feed/Feed";
 import Login from "./components/login/index";
 import Register from "./components/register/index";
 import Home from "./components/home/index";
@@ -109,6 +110,7 @@ class App extends Component {
               <Route exact path="/login" component={Login} />
               <Route exact path="/register" component={Register} />
               <Route exact path="/profile" component={Profile} />
+              <Route exact path="/feed" component={Feed} />
               <Route path="/user" component={BoardUser} />
               
             </Switch>
