+++ client/src/components/UserCard/UserCard.css
@@ -1,0 +1,17 @@
+.userCard .avatar {
+    display: block;
+    width:20%;
+    margin-right:0.5rem;
+}
+
+.userCard .userFlex {
+    display:flex;
+    flex-direction:row;
+    justify-content: flex-start;
+}
+
+.userCard h3 {
+    display:flex;
+    flex-direction: column;
+    justify-content: center;
+}
