+++ client/src/pages/Feed/Feed.js
@@ -1,15 +1,14 @@
 import React from "react";
 import Header from "../../components/Header/Header";
 import RecipeCard from "../../components/RecipeCard/RecipeCard";
-import "./style.css";
+import UserCard from "../../components/UserCard/UserCard"
+import "./Feed.css";
 import "../../App.css";
 
 export default function Feed() {
     return(
         <div id="Feed">
             <Header />
-
-            <div>Feed Page</div>
 
             <div className="feedWrap">
                 <div className="mainFeed">
@@ -18,12 +17,16 @@ export default function Feed() {
 
                 <div className="sidebarWrap">
                     <div className="popularRecipesFeed">
+                        <div className="sidebarHeader"></div>
                         <RecipeCard />
+                        <div className="sidebarFooter"></div>
                     </div>
 
-                    {/* <div className="popularUsersFeed">
+                    <div className="popularUsersFeed">
+                        <div className="sidebarHeader"></div>
                         <UserCard />
-                    </div> */}
+                        <div className="sidebarFooter"></div>
+                    </div>
                 </div>
             </div>
         </div>
